+++ XCredsLoginPlugIn/Mechanisms/XCredsLoginMechanism.swift
@@ -41,18 +41,7 @@ import Network
     @objc func tearDown() {
         TCSLogWithMark("Got teardown request")
 
-        let isAccountCreationPending = getHint(type: .isAccountCreationPending) as? Bool 
-
-        if isAccountCreationPending == false {
-            TCSLogWithMark("Hiding background")
-
-            for window in self.mainLoginWindowController!.windowArray{
-                window.close()
-            }
-        }
-        else {
-            TCSLogWithMark("Not hiding progress indicator to avoid black screen")
-        }
+     
     }
 
     override func reload() {